+++ functorch/functorch/_src/remat_utils_mincut.py
@@ -2,58 +2,16 @@ import torch
 from torch.fx.passes.infra.partitioner import CapabilityBasedPartitioner
 from torch.fx.passes.backends.nvfuser import NvFuserOperatorSupport
 from torch.fx.passes.tools_common import legalize_graph
-import torch.fx as fx
 import operator
 import math
 import copy
 
-from .utilities import _size_of
+from .utilities import _size_of, ban_recomputation
 
 
 num_group_remat = 0  # used for analytical purpose
 memory_reduced = 0
 num_node_pairs = 0
-
-
-aten = torch.ops.aten
-
-pointwise_ops = [aten.add, aten.sub, aten.div, aten.atan2, aten.mul, aten.max, aten.min, aten.pow, aten.remainder, aten.fmod, aten.__and__, aten.__or__, aten.__xor__, aten.__lshift__, aten.__rshift__, aten.eq, aten.ne, aten.ge, aten.gt, aten.le, aten.lt, aten.abs, aten.bitwise_not, aten.ceil, aten.floor, aten.frac, aten.neg, aten.relu, aten.round, aten.silu, aten.trunc, aten.log, aten.log10, aten.log1p, aten.log2, aten.lgamma, aten.exp, aten.expm1, aten.erf, aten.erfc, aten.cos, aten.acos, aten.cosh, aten.sin, aten.asin, aten.sinh, aten.tan, aten.atan, aten.tanh, aten.atanh, aten.sqrt, aten.rsqrt,  aten.reciprocal, aten.sigmoid, aten.softplus, aten.threshold, aten.threshold_backward, aten.clamp, aten.where, aten.lerp, aten.addcmul, aten.gelu, aten.gelu_backward]  # noqa: E501
-misc_ops = [aten.to, aten.type_as, operator.getitem]
-reduction_ops = [aten.softmax, aten._softmax, aten._softmax_backward_data, aten.sum, aten.mean, aten._grad_sum_to_size, aten.sum_to_size, aten.amax]  # noqa: E501
-
-
-norm_ops = [aten.instance_norm, aten._batch_norm_impl_index, aten.native_batch_norm, aten.batch_norm, aten._batch_norm_impl_index_backward, aten.native_layer_norm, aten.layer_norm, aten.native_layer_norm_backward]  # noqa: E501
-view_ops = [aten.squeeze, aten.unsqueeze]
-random_ops = [aten.native_dropout, aten.rand_like, aten.randn_like]
-compute_intensive_ops = [aten.mm, aten.convolution, aten.convolution_backward, aten.bmm, aten.addmm, aten.upsample_bilinear2d]  # noqa: E501
-unrecomputable_ops = random_ops + compute_intensive_ops + norm_ops
-
-recomputable_ops = set(
-    pointwise_ops
-    + misc_ops
-    + reduction_ops
-    + view_ops
-)
-fusible_ops = recomputable_ops | set(random_ops)
-
-AGGRESSIVE_RECOMPUTATION = False
-
-
-def ban_recomputation(node):
-    if AGGRESSIVE_RECOMPUTATION:
-        return (node.op == 'call_function' and node.target in unrecomputable_ops)
-    else:
-        if node.op != 'call_function':
-            return False
-        if node.target not in recomputable_ops:
-            return True
-        # If the output of the reduction is 4x smaller (arbitrary choice),
-        # then we don't allow recomputation.
-        if node.target in reduction_ops:
-            input_tensors_size = sum(_size_of(i.meta['tensor_meta']) for i in node.args if isinstance(i, fx.Node))
-            output_size = _size_of(node.meta['tensor_meta'])
-            return (output_size * 4 < input_tensors_size)
-        return False
 
 
 def is_fused_node(node):
@@ -196,7 +154,7 @@ def add_new_outputs(node_pair, fused_graph, name_to_node, module_origin, cut_nod
                             operator.getitem, args=(node_pair[0], i + len(old_args), ))
                 new_args = list(old_args) + module_origin_new_outputs
                 module_origin.graph.erase_node(node)
-                module_origin.graph.output(new_args[0] if len(new_args) == 1 else tuple(new_args))       
+                module_origin.graph.output(new_args[0] if len(new_args) == 1 else tuple(new_args))
             break
     module_origin.recompile()
     fused_graph.recompile()
@@ -383,7 +341,7 @@ def find_min_cut(node_pair, node_users_map, fused_graph):
 
         read_cost = weight
 
-        capacity = write_cost+read_cost
+        capacity = write_cost + read_cost
         return capacity
 
     for node in module_origin.graph.nodes:
@@ -393,7 +351,7 @@ def find_min_cut(node_pair, node_users_map, fused_graph):
         weight = get_weight(node)
 
         if ban_recomputation(node):
-            nx_graph.add_edge("source",  node.name+"_out", capacity=math.inf)
+            nx_graph.add_edge("source", node.name + "_out", capacity=math.inf)
 
         # some ops like cuda_batch_norm return tuples, and they cannot be returned as output
         # because torch.jit.script does not accept
@@ -406,18 +364,18 @@ def find_min_cut(node_pair, node_users_map, fused_graph):
 
         if node.op == 'placeholder':
             capacity = weight
-            nx_graph.add_edge("source", node.name+"_in", capacity=math.inf)
+            nx_graph.add_edge("source", node.name + "_in", capacity=math.inf)
         elif node.op == 'call_function':
             capacity = get_capacity(node)
 
         if (node.name in dest_placeholder_names or
            (node.name in getitem_users and getitem_users[node.name] in dest_placeholder_names)):
-            nx_graph.add_edge(node.name+"_out", 'sink', capacity=capacity)
+            nx_graph.add_edge(node.name + "_out", 'sink', capacity=capacity)
 
-        nx_graph.add_edge(node.name+"_in", node.name+"_out", capacity=capacity)
+        nx_graph.add_edge(node.name + "_in", node.name + "_out", capacity=capacity)
         for user in node.users:
             if user.op != "output":
-                nx_graph.add_edge(node.name+"_out", user.name+"_in", capacity=math.inf)
+                nx_graph.add_edge(node.name + "_out", user.name + "_in", capacity=math.inf)
 
     cut_value, partition = nx.minimum_cut(nx_graph, "source", "sink")
